migrate_listings

+++ database/migrations/2022_04_22_163902_create_listings_table.php
@@ -13,7 +13,9 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('listings', function (Blueprint $table) {
+        if (!Schema::hasTable('listings')) {
+            Schema::create('listings', function (Blueprint $table)
+        {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->string('title');
@@ -27,6 +29,7 @@ return new class extends Migration
             $table->timestamps();
         });
     }
+}
 
     /**
      * Reverse the migrations.
